decoupling ZeroMQ pub/sub component into separate module
